Transfer voodoo damage/item pickup to the main player object
While vanilla behavior is fine in most cases, when reset level on death is
disabled, it tends to cause problems. Making a "fake" player object type and
transferring any attempts to damage it, or item pickups it does, to the main
player object makes actions like the TNT MAP30 puzzle insta-kill repeatable.